+++ config/config.go
@@ -10,8 +10,11 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"go.opentelemetry.io/otel/log"
+	nooplog "go.opentelemetry.io/otel/log/noop"
 	"go.opentelemetry.io/otel/metric"
+	noopmetric "go.opentelemetry.io/otel/metric/noop"
 	"go.opentelemetry.io/otel/trace"
+	nooptrace "go.opentelemetry.io/otel/trace/noop"
 )
 
 const (
@@ -62,31 +65,41 @@ func (s *SDK) Shutdown(ctx context.Context) error {
 	return s.shutdown(ctx)
 }
 
+var noopSDK = SDK{
+	loggerProvider: nooplog.LoggerProvider{},
+	meterProvider:  noopmetric.MeterProvider{},
+	tracerProvider: nooptrace.TracerProvider{},
+	shutdown:       func(ctx context.Context) error { return nil },
+}
+
 // NewSDK creates SDK providers based on the configuration model.
 func NewSDK(opts ...ConfigurationOption) (SDK, error) {
 	o := configOptions{}
 	for _, opt := range opts {
 		o = opt.apply(o)
 	}
+	if o.opentelemetryConfig.Disabled != nil && *o.opentelemetryConfig.Disabled {
+		return noopSDK, nil
+	}
 
 	r, err := newResource(o.opentelemetryConfig.Resource)
 	if err != nil {
-		return SDK{}, err
+		return noopSDK, err
 	}
 
 	mp, mpShutdown, err := meterProvider(o, r)
 	if err != nil {
-		return SDK{}, err
+		return noopSDK, err
 	}
 
 	tp, tpShutdown, err := tracerProvider(o, r)
 	if err != nil {
-		return SDK{}, err
+		return noopSDK, err
 	}
 
 	lp, lpShutdown, err := loggerProvider(o, r)
 	if err != nil {
-		return SDK{}, err
+		return noopSDK, err
 	}
 
 	return SDK{